refactor(perl-parser-core): split cpan_pattern_tests.rs into 16 standalone test files
Split the monolithic 1,642-line cpan_pattern_tests.rs (172 tests, 16 modules)
into 16 standalone test files with a shared helper module. This prevents merge
conflicts when multiple agents add CPAN pattern tests simultaneously.

New structure:
- tests/cpan_test_helpers/mod.rs — shared parse/assert_clean_parse/top_level_kinds
- tests/cpan_moose_moo.rs (17 tests)
- tests/cpan_try_tiny.rs (6 tests)
- tests/cpan_list_util.rs (9 tests)
- tests/cpan_dbi.rs (9 tests)
- tests/cpan_file_find.rs (5 tests)
- tests/cpan_oo_patterns.rs (11 tests)
- tests/cpan_data_structures.rs (11 tests)
- tests/cpan_functional.rs (11 tests)
- tests/cpan_error_handling.rs (8 tests)
- tests/cpan_regex_patterns.rs (9 tests)
- tests/cpan_module_patterns.rs (11 tests)
- tests/cpan_statement_modifiers.rs (8 tests)
- tests/cpan_variable_declarations.rs (12 tests)
- tests/cpan_real_world_programs.rs (8 tests)
- tests/cpan_misc_idioms.rs (28 tests)
- tests/cpan_compound_stmt_modifier.rs (9 tests)

Also fixes CLAUDE.md test template which referenced nonexistent
perl_parser_core::parse function.
